adding Math and Physic class

+++ Console Game/Engine/Engine.cpp
@@ -7,7 +7,7 @@
 int Engine::SCREEN_WIDTH = 1024;
 int Engine::SCREEN_HEIGHT = 768;
 GLFWwindow* Engine::window = NULL;
-double Engine::dt = 0;
+float Engine::dt = 0;
 
 Engine::Engine()
 {
@@ -65,14 +65,14 @@ bool Engine::Initialize(const char* windowTitle) // const added
 	glEnable(GL_BLEND);
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
-	lastTime = glfwGetTime();
+	lastTime = (float)glfwGetTime();
 
 	return true;
 }
 
 void Engine::Update()
 {
-	double now = glfwGetTime();
+	float now = (float)glfwGetTime();
 	dt = (now - lastTime);
 	lastTime = now;
 	glfwPollEvents(); // handles event in queue
@@ -93,7 +93,7 @@ void Engine::EndRender(){
 }
 
 
-double Engine::GetDT() {
+float Engine::GetDT() {
 
 	return dt;
 }
